More compatibility and stability updates for macOS Sonoma 14.x: - Adjusted code to deal with the majority of 'loss of precision' warnings. - Modified and modernized the application windows so the widgets sized accordingly and didn't show haphazardly on the screen. - Updated the deprecated keychain interface to use newer search, add, and update routines. - Updated the deprecated calendar date routines to use date formatters. - Fixed interface to use SSL port 993 as default, which is mostly the case anymore.

+++ comms_ssl.c
@@ -144,7 +144,7 @@ size_t comms_ssl_read( void *buf, size_t count )
 	}
 
 	do {
-		res = SSL_read(ssl, buf, count);
+		res = SSL_read(ssl, buf, (int)count);
 		if ( res <= 0 ) {
 			err = SSL_get_error(ssl, res);
 			alert("res: %d, err: %d, errno: %d\n", res, err, errno );
@@ -167,7 +167,7 @@ size_t comms_ssl_write( void *buf, size_t count )
 	}
 
 	do {
-		res = SSL_write(ssl, buf, count);
+		res = SSL_write(ssl, buf, (int)count);
 		if ( res <= 0 ) {
 			err = SSL_get_error(ssl, res);
 		}
